requirements.txt
# Веб-сервис
fastapi==0.115.5
uvicorn[standard]==0.32.0

# Модели и обучение
torch==2.3.1
transformers==4.45.2

# Работа с данными и PDF
pandas==2.2.3
pdfminer.six==20231228

# Схемы и загрузка файлов
pydantic==2.9.2
python-multipart==0.0.12
python-docx==1.1.2